requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tiktoken>=0.5.1
google-generativeai>=0.7.0
python-docx>=1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
numpy>=1.24.0
python-dotenv>=1.0.1

